feat: /setup 관련 layer 엔티티, 컨트롤러 추가

+++ workspace/server/src/main/java/org/hellokicktty/server/controller/KickboardController.java
@@ -1,7 +1,6 @@
 package org.hellokicktty.server.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.hellokicktty.server.domain.Kickboard;
 import org.hellokicktty.server.dto.KickboardResponseDto;
 import org.hellokicktty.server.service.KickboardService;
 import org.springframework.web.bind.annotation.*;
@@ -17,30 +16,31 @@ public class KickboardController {
     private final KickboardService kickboardService;
 
     @GetMapping
-    List<Kickboard> findKickboardsInRange(Double lat, Double lng) {
+    List<KickboardResponseDto> findKickboardsInRange(Double lat, Double lng) {
         List<KickboardResponseDto> dto = new ArrayList<>();
         kickboardService.findKickboardsInRange(lat, lng)
                 .stream()
                 .forEach(item -> dto.add(new KickboardResponseDto(item)));
-        return kickboardService.findKickboardsInRange(lat, lng);
+        return dto;
     }
 
     @GetMapping("/{id}")
-    Kickboard findKickboard(@PathVariable Long id) {
-        return kickboardService.findKickboard(id);
+    KickboardResponseDto findKickboard(@PathVariable Long id) {
+        return new KickboardResponseDto(kickboardService.findKickboard(id));
     }
 
 
     @PostMapping
-    Kickboard addKickboard(Long id, Double lat, Double lng) {
+    KickboardResponseDto addKickboard(Long id, Double lat, Double lng) {
         kickboardService.addKickboard(id, lat, lng);
-        return kickboardService.findKickboard(id);
+        return new KickboardResponseDto(kickboardService.findKickboard(id));
     }
+
 
     @DeleteMapping("/{id}")
-    void addKickboard(@PathVariable Long id) {
+    void removeKickboard(@PathVariable Long id) {
+        if (kickboardService.findKickboard(id) == null) return;
         kickboardService.removeKickboard(id);
     }
-
 
 }
